cloud_function/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
openai==0.27.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-api-core==2.11.0
Jinja2==3.1.2
gunicorn==20.1.0
Markdown==3.4.3
MarkupSafe==2.1.2